+++ src/codegen/mod.rs
@@ -11,12 +11,8 @@ use sema::ast::Namespace;
 /// not all contracts need a cfg; only those for which we need the
 pub fn codegen(contract_no: usize, ns: &mut Namespace) {
     if ns.contracts[contract_no].is_concrete() {
-        // we need to iterate over the contracts function table, while generating cfg for each. We can't
-        // iterate mutably over the function table because then we cannot borrow the namespace for generating
-        // the cfg. So first populate a separate hash and then set in another loop
-        let mut generated = Vec::new();
-
         let mut cfg_no = 0;
+        let mut all_cfg = Vec::new();
 
         // all the functions should have a cfg_no assigned, so we can generate call instructions to the correct function
         for (_, func_cfg) in ns.contracts[contract_no].all_functions.iter_mut() {
@@ -24,35 +20,46 @@ pub fn codegen(contract_no: usize, ns: &mut Namespace) {
             cfg_no += 1;
         }
 
-        ns.contracts[contract_no]
-            .cfg
-            .resize(cfg_no, ControlFlowGraph::placeholder());
+        all_cfg.resize(cfg_no, ControlFlowGraph::placeholder());
 
-        for ((base_contract_no, function_no), cfg_no) in &ns.contracts[contract_no].all_functions {
-            let c = cfg::generate_cfg(contract_no, *base_contract_no, Some(*function_no), ns);
-            generated.push((*cfg_no, c));
-        }
-
-        for (cfg_no, cfg) in generated.into_iter() {
-            ns.contracts[contract_no].cfg[cfg_no] = cfg;
+        // clone all_functions so we can pass a mutable reference to generate_cfg
+        for (base_contract_no, function_no, cfg_no) in ns.contracts[contract_no]
+            .all_functions
+            .iter()
+            .map(|((base_contract_no, function_no), cfg_no)| {
+                (*base_contract_no, *function_no, *cfg_no)
+            })
+            .collect::<Vec<(usize, usize, usize)>>()
+            .into_iter()
+        {
+            cfg::generate_cfg(
+                contract_no,
+                base_contract_no,
+                Some(function_no),
+                cfg_no,
+                &mut all_cfg,
+                ns,
+            )
         }
 
         // Generate cfg for storage initializers
         let cfg = storage_initializer(contract_no, ns);
-        let pos = ns.contracts[contract_no].cfg.len();
-        ns.contracts[contract_no].cfg.push(cfg);
+        let pos = all_cfg.len();
+        all_cfg.push(cfg);
         ns.contracts[contract_no].initializer = Some(pos);
 
         if !ns.contracts[contract_no].have_constructor() {
             // generate the default constructor
             let func = ns.default_constructor();
+            let cfg_no = all_cfg.len();
+            all_cfg.push(ControlFlowGraph::placeholder());
 
-            let pos = ns.contracts[contract_no].cfg.len();
-            let cfg = cfg::generate_cfg(contract_no, contract_no, None, ns);
-            ns.contracts[contract_no].cfg.push(cfg);
+            cfg::generate_cfg(contract_no, contract_no, None, cfg_no, &mut all_cfg, ns);
 
-            ns.contracts[contract_no].default_constructor = Some((func, pos));
+            ns.contracts[contract_no].default_constructor = Some((func, cfg_no));
         }
+
+        ns.contracts[contract_no].cfg = all_cfg;
     }
 }