+++ src/sema/printer.rs
@@ -445,6 +445,7 @@ fn print_statement(stmts: &[Statement], func: &Function, ns: &Namespace) -> Vec<
 
                 Tree::Branch(String::from("try-catch"), list)
             }
+            Statement::Underscore(_) => Tree::Leaf(String::from("underscore")),
         });
     }